don't make an API call to LandingPage.loadContent() in testing
it causes the API to hang while waiting for an API response that doesn't arrive

```
 isAxiosError: true,
        toJSON: [Function] }".
      at BufferedConsole.log (node_modules/@jest/console/build/BufferedConsole.js:199:10)
      at LandingPage.loadContent (src/scenes/LandingPage/LandingPage.js:91:17)

(node:492) UnhandledPromiseRejectionWarning: TypeError: Cannot read property 'body' of null
(node:492) UnhandledPromiseRejectionWarning: Unhandled promise rejection. This error originated either by throwing inside of an async function without a catch block, or by rejecting a promise which was not handled with .catch(). (rejection id: 2)
(node:492) [DEP0018] DeprecationWarning: Unhandled promise rejections are deprecated. In the future, promise rejections that are not handled will terminate the Node.js process with a non-zero exit code.

  ●  Cannot log after tests are done. Did you forget to wait for something async in your test?
    Attempted to log "{ Error: connect ECONNREFUSED 127.0.0.1:8000
        at TCPConnectWrap.afterConnect [as oncomplete] (net.js:1106:14)
      errno: 'ECONNREFUSED',
      code: 'ECONNREFUSED',
      syscall: 'connect',
      address: '127.0.0.1',
      port: 8000,
      config:
       { url: 'http://127.0.0.1:8000/api/search/?q=Ontario',

```