disable application score input after winner has been selected
We are only making this change in the frontend
and not changing in the backend in case there is a use case
where we need to modify the score.
So we don't want to make it completely blocked,
but rather, just add some friction so that if we are changing anything,
it's being done deliberately and with intent.